+++ frb_rust/src/lockable/mod.rs
@@ -3,3 +3,4 @@ mod immediate_invalid_borrow_checker;
 pub(crate) mod implementor;
 mod order;
 mod order_info;
+mod order_computer;
